Add Camera and CameraDecorator classes
Lots of the meat I stole from learnopengl, but I decided to split the Camera there into a Camera and a wrapper that gives extra ability to manipulate the Camera vectors i.e. the CameraDecorator. The interface from the decorator very much belies its concrete source which is the yaw-pitch driven camera in learnopengl. Meh.